Send email for CI results
Summary:
BPF CI has traditionally been just providing a hint to maintainers about the
status of a patch. However, it has taken on more and more of a prominent role as
part of the patch lifecycle in recent times (~years).
Currently there is no notification mechanism available that informs patch
submitters (or maintainers, for that matter) of the status of a patch. That's
obviously not great, as it requires users to poll for the outcome.
With this change we introduce such a mechanism, by adjusting KPD to send emails
when CI results come in. We don't have BPF CI baseline tracking (to decide
whether the upstream branch is in good shape or not). Instead, we ensure to not
spam users by keeping track of the CI state and only send an email if this state
changes (i.e., pass -> fail or vice versa). We use GitHub PR labels to maintain
that state, to not have to worry about KPD crashes and their impact too much.

Currently emails will only be sent to a (configurable but) fixed set of email
addresses. That is what we intend to do for validation purposes. Once we have
confidence in the logic we can adjust it to include the patch submitter as well.

Reviewed By: chantra

Differential Revision: D48531436

fbshipit-source-id: 07be9b815da920e0599ee96fa6131e04dc2e86df