Introduce email notification submitter allowlist
Summary:
We want to onboard Meta-internal kernel developers to the email notification
workflow, by sending them emails about CI workflow run results of their own
submissions.
This change puts the necessary infrastructure in place. Specifically, we
introduce the 'submitter_allowlist' attribute to the configuration, which is
meant to contain a list of email addresses. If a CI result comes in and the
submitters email address is in this list an email will be sent to said submitter
(in addition to all the recipients that we send emails to unconditionally).

Reviewed By: chantra

Differential Revision: D49711217

fbshipit-source-id: be2cd0be34d703d4185ac164e080a90efc526507