Use workflows runs instead of check suites
Summary:
Our sync_checks() logic heavily relies on what GitHub refers to as check suites.
They are non-obvious and underdocumented concept that does not map cleanly to
anything that users (developers) interface with. Rather, what users work with
are pull requests and pull requests may trigger workflow runs (workflow runs are
based on workflow definitions, which is what developers write).
This divergence makes debugging difficult. Certainly when navigating from a pull
request (from the GitHub UI), it's utterly unclear how we'd ever arrive at a
check suite. Perhaps it's a UI exposed concept, but I haven't found the
interface.
There is some relationship between workflow runs and check suites, but it's not
entirely clear what it is and how to reason about it, let alone build a mental
model.
In short, check suites are a questionable concept that are unnecessarily hard to
reason about for our intents and purposes.
This change proposes getting rid of their usage and just working with workflows
instead. We already need to iterate workflows anyway, because, you guessed it,
there is no way to infer the URL of a workflow run from a check suite; but we
very much need this URL for error reporting purposes.

Reviewed By: chantra

Differential Revision: D49711218

fbshipit-source-id: 20060ebe44160517b2f1cf047c4a052f451fbe1b