Make Cc: configurable in email configuration
Summary:
The intention of sending BPF CI email notifications is first and foremost to
inform patch submitters about their patch's status. But on top of that BPF
maintainers should be notified. To make it more clear that action is ultimately
on the patch submitter, it makes sense to address the email to them (To:) and
have maintainers be copied (Cc:).
So far we only supported To: addressing. With this change we make the Cc:
configurable as well.

Reviewed By: yurinnick

Differential Revision: D49834802

fbshipit-source-id: 28f060d09601d3c7b47adb5232636fb4cfadb514