Thread CI emails into upstream patch set email thread
Summary:
We have received feedback that it would be useful to thread CI emails into the
original patch submission email thread, so that it is easier to correlate emails
logically belonging together.
Set the In-Reply-To header accordingly when sending our emails. Because we now
thread into upstream as opposed to acting as our own thread root, revise the
message ID creation accordingly.

Reviewed By: chantra

Differential Revision: D50035627

fbshipit-source-id: 0970507eb4a1203c6400dcaad7790cc43dcfabb8