Work with proper Series objects
Summary:
In the "relevant subject retrieval" code we retrieves all patches fulfilling
certain criteria and then butcher together some manifestation of a Series
object based on chosen contents of the response. On other paths we actually
retrieve the proper Series API object.
It's confusing to have something representing a Patchwork Series that isn't
actually created from a proper Patchwork API Series object sometimes (and,
hence, matching the corresponding documentation). It also has bitten us already
in the past in D49711217.
With this change we retrieve the proper Series object for each patch, to
mitigate this issue. The get_series_by_id() call checks for previously cached
series and so we are unlikely to hit any request limits as part of this work.

Reviewed By: chantra, yurinnick

Differential Revision: D50035628

fbshipit-source-id: 7d3faf9745bec67ca8f76859a186df7d33426ccf