Highlight CI status more prominently
Summary:
Some email clients may not show the email's subject all that visibly, making it
hard to rapidly infer the status of a run based on only the email body contents.
Highlight the status some more by mentioning it at the beginning of the email to
allow for quicker inference of the outcome.

Reviewed By: yurinnick

Differential Revision: D50141733

fbshipit-source-id: 03f6478ed2c81d57c6e3d72ab7d487b908001913