Evaluate not just complete runs in sync_checks()
Summary:
As part of D49711218 we changed the workflow runs that we look at from all to
only complete ones. It intuitively makes sense, but that's not actually correct,
because we use the knowledge that there are outstanding runs when determining
overall workflow status: if pending runs are present we do not yet report status
(unless others failed, in which case we can never recover to "success" and might
as well report failure early).
This change makes us go back to evaluating all runs and their conclusions.

Reviewed By: chantra

Differential Revision: D50335750

fbshipit-source-id: abdf06905f53561309e773df21a66800151dac62