Handle workflow cancellation more gracefully
Summary:
When KPD rebases patch sets due to upstream changes of the base branch, it force
pushes the now-rebased patch to trigger a re-test. Because of how our workflows
are configured, that will cancel any existing runs in the process. However, the
GitHub API just reports a failure of the original, canceled run.
With this change we attempt to detect such cancellations and treat them as
"pending" for the time being, as we do for runs that don't have a result yet.

Reviewed By: yurinnick

Differential Revision: D50344386

fbshipit-source-id: f765ee410ad8083398c07a08a20dcb74800f36e1