Handle interrupted workflow runs more gracefully
Summary:
CI run 6656321967 [0] is an example of a run that got interrupted because the
runner died, lost connectivity, or similar. We do not want to report a failure
for such infrastructure issues, as they do not reflect a test result of the
actual patch.
With this change we attempt to detect such failures and map them to a "pending"
state, until we eventually retry.

[0] https://github.com/kernel-patches/bpf/actions/runs/6656321967

Reviewed By: chantra

Differential Revision: D50701365

fbshipit-source-id: 4034c77d5f516ea8fcf54c93c60b7495b1ecd7b4