Make sure to work on up-to-date PR data
Summary:
As laid out in T168827021, we are prone to working with outdated PR data,
because the Python GitHub module we use caches state instead of re-requesting
it. This can be a source of trouble. In the above task it causes us to associate
GitHub CI Workflow runs for an outdated SHA-1 with an updated patchwork
series/pull request.
With this change we force an update of the pull request data to make sure that
`pr.head.sha` is up-to-date. Note that there really doesn't seem to be a single
best located for requesting such an update, so I am adding it where it's needed
at this point.

Reviewed By: anakryiko

Differential Revision: D50910044

fbshipit-source-id: e3ec61a4cb202917be1b73c1b372648f55ee054f