Handle PR closure more gracefully
Summary:
When we close a PR it we just mark it as closed. Subsequently, if it pops up
again for evaluation we end up returning None from checkout_and_patch, because,
well, it special cases closed PRs and returns None. It is wrong to assert that
the PullRequest object returned by said function is not None without having
taken any steps to prevent closed pull request from being closed at this point.
To fix issue, just ignore closed pull requests at this point by moving to the
next patch/series.

Reviewed By: yurinnick

Differential Revision: D51165509

fbshipit-source-id: 5605052c2a6377f90542cae7992c9cef210a16bd