Reference CI base object in email body
Summary:
Currently we always use the Patchwork series' name in the body of our CI emails.
That will become confusing once we adjust the subject of said emails to reflect
the cover letter OR last patch, as per the revised logic.
With this change we use the previously introduced get_ci_base() function to
retrieve the name that we will also use in the subject moving forward, to
prevent any confusion.

Reviewed By: yurinnick

Differential Revision: D51165510

fbshipit-source-id: 061ae19ff93ea5b09eb62596ab12a2e444b4c702