Use upstream conversation subject for CI emails
Summary:
We have had requests to use the subject of the upstream discussion when sending
BPF CI email messages. The reason being that Gmail doesn't thread emails
properly once the subject changes [0] -- and so our indication of
failure/success and other goodies in the subject cause it to keep any CI emails
separate from the main discussion.
With this change we switch over to using the upstream subject, similar to what
bots such as kernel test robot do [1].

[0] https://support.google.com/mail/answer/5900
[1] https://lore.kernel.org/all/ZRkI4j2+dXZ9sjQ4@f61ccfac960a/

Reviewed By: chantra

Differential Revision: D51139610

fbshipit-source-id: 63aa1bc27bfc5493ded14f7c1498822186b1490a