Send emails in response to last patch
Summary:
Some submissions, such as
https://lore.kernel.org/all/20231108051430.1830950-1-andrii@kernel.org/, despite
containing multiple patches, evidently do not come with a cover letter.
If no cover letter is present we choose to send CI emails in response to the
first patch. However, given that we test the last patch it makes more sense to
also send the email as a response to the last patch. So use its message ID.

Reviewed By: yurinnick

Differential Revision: D51165507

fbshipit-source-id: 70c92c943f0151fd7200837544d8dca1b2407fc9