branch_worker: Check pr.user.login for relevant pr (#14)
Summary:
When a PR is created from a different branch (instead of a fork), we see pr['head']'user']['login'] and pr['base']['user']['login'] are both "kernel-patches", but pr['user']['login'] is different. Such PR should not be considered as relevant. Add a check to cover this case.

Pull Request resolved: https://github.com/facebookincubator/kernel-patches-daemon/pull/14

Reviewed By: liu-song-6

Differential Revision: D52885119

Pulled By: chantra

fbshipit-source-id: 1e15c48d1ac82f72c8680473cf895a097f72048f