Add more logs around PR tracking
Summary:
It looks like KPD gets confused when patchwork
changes the name of the series. The name of the series
can change when not all emails in the set are received
by patchwork at the same time.

It looks like `_guess_pr` contains logic to detect PR
title changes, but for some reason it isn't working.

Add more logs to troubleshoot this.

Reviewed By: chantra

Differential Revision: D57278464

fbshipit-source-id: 8fa926f5084dec159ded1602735698469279869b