Turn submitter_allowlist into a list of regex
Summary:
Previously, the allowlist was for exact string matches. For a more
gradual and controlled rollout, we want regex.

What I have in mind is gating a rollout by the first letter of the submitter
email.  So for example, the first phase would be letters [a-fA-F] or something.
This makes it easy to tell which submitters should be getting emails.

Differential Revision: D58474179

fbshipit-source-id: 64cbe5cc74784559d425114b18d59b5af2151490