Validate that all tag/branches mapping exist
Summary:
When matching a tag to a branch, kpd will error out if the branch does not
exists.
This error used to happen after cloning and after we had gone through some
iterations, and could possibly not be triggered until a patch with a specific
tag makes it to the mailing list.

To avoid breaking out of the blue based on input, let's validate the config and
crash then so it can be fixed early.

Reviewed By: danobi

Differential Revision: D61306581

fbshipit-source-id: 97aa2a2edbe579dbe2b0c096e6cdc8c94d5252af