build: Remove poetry.lock from source control
We have had a bunch of PRs to tend to due to dependabot:
https://github.com/facebookincubator/kernel-patches-daemon/pulls?q=is%3Apr++label%3Adependencies+

In most cases, those were only changing the depedency in the .lock file.
We are not getting much value of shipping this file, and will be better off
by getting rid of it and fixign the remotely possible breakage in the future
vs dealing with those weekly.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>