build: Remove poetry.lock from source control (#59)
Summary:
We have had a bunch of PRs to tend to due to dependabot: https://github.com/facebookincubator/kernel-patches-daemon/pulls?q=is%3Apr++label%3Adependencies+

In most cases, those were only changing the depedency in the .lock file. We are not getting much value of shipping this file, and will be better off by getting rid of it and fixign the remotely possible breakage in the future vs dealing with those weekly.

Pull Request resolved: https://github.com/facebookincubator/kernel-patches-daemon/pull/59

Reviewed By: chantra

Differential Revision: D62477549

fbshipit-source-id: 6651931ef74bfbf58a4ef5762dcbbb1ed3e39c63