email: update cc list to enable multiple recipients
Actually, I was getting an error with just one email.  Further below
From my change on this LOC:

	for to in to_list + cc_list:

I was getting an error about concatenating a list and a string.  The to_list
Was clearly a list based on how it was processed just above the for loop.

The cc_list printed its type as a list but it looked like a string logging it.

Either way, single or multiple entry cc in the config file, this patch works.

Signed-off-by: Paul Luse <paul.e.luse@intel.com>